+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Show cluster namespace if available in the metric
+- Add success message if no test clusters are left over
 
 ## [1.4.0] - 2024-04-09